Only test POSIX consistency against JVM
A test was failing in JS due to differences in how the `\p{ASCII}`
character class behaves.
